(refactor) examples
